requirements.txt
Django~=3.2.8
django-model-utils~=4.1.1
django-admin-honeypot~=1.1.0
python-environ~=0.4.54
